refactor(console): extract workspace details tray from context tray
Move the plumbing status-row rendering (Storage/Sync/File tools/Server
handoff/Handoff list/ACP handoff) out of ConsoleWorkspaceContextTray into
a new ConsoleWorkspaceDetailsTray, and add show_heading to the context
tray so Task 9 can mount both trays side by side in the left rail. Moved
members are byte-identical; only the host class changed.

chat_screen.py is untouched (task 9's job), so the Console screen
currently mounts only the smaller context tray. This exposes a
pre-existing, test-harness-only layout bug (unstyled 1fr Horizontal
wrappers in the conversation browser, masked in production by real CSS
that pins their height) that now clips conversation rows out of the
clickable rail viewport in some screen-level tests. Deferring the
following 23 tests to task 9, per the plan's sanctioned sequencing gap
(root-caused and cross-checked via git-stash A/B testing and an
independent research pass; see .superpowers/sdd/task-7-report.md):

Direct on-screen-presence-of-moved-rows assertions:
- test_console_workspace_context_rail.py::test_console_workspace_context_keeps_status_rows_below_grouped_browser
- test_console_workspace_context_rail.py::test_console_workspace_many_conversations_keep_lower_status_reachable
- test_console_workspace_context_rail.py::test_console_workspace_context_exposes_new_conversation_for_default_workspace
- test_console_workspace_context_rail.py::test_console_workspace_context_renders_active_workspace
- test_console_workspace_context_rail.py::test_console_workspace_context_renders_server_readiness_handoff_and_acp_contracts
- test_console_native_chat_flow.py::test_console_workspace_authority_rows_are_structured_for_scanning
- test_console_native_chat_flow.py::test_console_workspace_rail_new_conversation_creates_default_workspace_session

Indirect fallout (same root cause, click/text-wait misses):
- test_console_native_chat_flow.py::test_console_workspace_conversation_switch_restores_transcript_messages
- test_console_native_chat_flow.py::test_console_conversation_browser_lists_all_workspace_groups
- test_console_native_chat_flow.py::test_console_browser_selecting_non_default_workspace_native_session_switches_active_workspace
- test_console_native_chat_flow.py::test_console_browser_selecting_non_default_workspace_persisted_row_switches_active_workspace_before_resume
- test_console_native_chat_flow.py::test_console_browser_selecting_default_native_session_switches_to_default_and_keeps_file_tools_disabled
- test_console_native_chat_flow.py::test_console_browser_selecting_default_persisted_row_switches_to_default_and_keeps_file_tools_disabled
- test_console_native_chat_flow.py::test_console_browser_selecting_global_persisted_row_preserves_active_workspace
- test_console_native_chat_flow.py::test_console_conversation_browser_marks_unavailable_keeps_browsing_enabled
- test_console_native_chat_flow.py::test_console_conversation_browser_long_list_keeps_readiness_rows_reachable
- test_console_native_chat_flow.py::test_console_workspace_conversation_search_selection_keeps_query_active
- test_console_native_chat_flow.py::test_console_workspace_conversation_search_blank_selection_keeps_composer_focus
- test_console_native_chat_flow.py::test_console_workspace_conversation_search_selection_invalidates_pending_worker
- test_console_native_chat_flow.py::test_console_workspace_conversation_row_switches_native_session
- test_console_native_chat_flow.py::test_console_workspace_conversation_row_resumes_persisted_conversation
- test_console_native_chat_flow.py::test_console_workspace_conversation_resume_uses_persisted_workspace
- test_console_native_chat_flow.py::test_console_workspace_conversation_resume_uses_real_local_services

Also repointed 4 pre-existing tests that called the moved static methods
directly on ConsoleWorkspaceContextTray (_friendly_status_label,
_friendly_detail_copy, _split_status_row) at ConsoleWorkspaceDetailsTray
instead, since these assert unit behavior, not on-screen presence.
